# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pre-Autorización Quirúrgica",
    "description": "API del agente de pre-autorización quirúrgica en tiempo real. Recibe informes médicos + pólizas, valida cobertura/carencia/docs, y emite pre-aprobación, solicitud de docs faltantes o escalamiento al auditor médico humano. NUNCA auto-rechaza por diseño.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cases": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Submit a new authorization case",
        "description": "Crea un caso de autorización y dispara el flow de decisión.\n\nEn MVP el flow MOCK (rule engine determinístico) corre síncrono y\ndevuelve el caso decidido. En producción (B3) se devolvería 202 con\n`status=PENDIENTE` y la decisión llegaría por SSE.",
        "operationId": "submit_case_api_v1_cases_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseSubmitIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "List authorization cases",
        "description": "Lista paginada de casos visibles para el rol del usuario.\n\nEl rol se extrae del JWT (no del query string) para evitar escalada de\nprivilegios — un hospital nunca puede pedir vista de auditor.",
        "operationId": "list_cases_api_v1_cases_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro por status del caso.",
              "title": "Status"
            },
            "description": "Filtro por status del caso."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseOut"
                  },
                  "title": "Response List Cases Api V1 Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/upload": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Submit a new authorization case with a PDF report (multipart)",
        "description": "PDF upload variant of `POST /api/v1/cases`. Stores the file via FileStorage.",
        "operationId": "submit_case_upload_api_v1_cases_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_case_upload_api_v1_cases_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Get a case by id",
        "description": "Detalle de un caso. 404 si no existe.",
        "operationId": "get_case_api_v1_cases__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/trace": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Get the agent trace for a case",
        "description": "Traza completa del agente para el caso. 404 si no existe.",
        "operationId": "get_case_trace_api_v1_cases__case_id__trace_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseTraceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/resolve": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Resolve an escalated case (auditor only)",
        "description": "El auditor cierra un caso escalado.\n\nSolo accesible para `Role.AUDITOR`. La decisión queda registrada con\n`decided_by=AUDITOR`, `confidence=1.0`, `outcome=DECIDED`.",
        "operationId": "resolve_case_api_v1_cases__case_id__resolve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveCaseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/files/{filename}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Download a file attached to a case (auth + RBAC)",
        "description": "Serves a stored PDF after RBAC + path-traversal checks.",
        "operationId": "download_case_file_api_v1_cases__case_id__files__filename__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_submit_case_upload_api_v1_cases_upload_post": {
        "properties": {
          "policy_number": {
            "type": "string",
            "title": "Policy Number"
          },
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "procedure_solicited_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Procedure Solicited Hint"
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "attending_doctor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attending Doctor"
          },
          "scenario_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Key"
          }
        },
        "type": "object",
        "required": [
          "policy_number",
          "patient_id",
          "file"
        ],
        "title": "Body_submit_case_upload_api_v1_cases_upload_post"
      },
      "CaseLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self"
          },
          "trace": {
            "type": "string",
            "title": "Trace"
          },
          "resolve": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolve"
          },
          "events": {
            "type": "string",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "self",
          "trace",
          "events"
        ],
        "title": "CaseLinks",
        "description": "HATEOAS-lite — paths relativos para el cliente."
      },
      "CaseOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "reportId": {
            "type": "string",
            "title": "Reportid"
          },
          "policyNumber": {
            "type": "string",
            "title": "Policynumber"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "decidedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decidedat"
          },
          "_links": {
            "$ref": "#/components/schemas/CaseLinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "reportId",
          "policyNumber",
          "createdAt",
          "_links"
        ],
        "title": "CaseOut",
        "description": "Caso de autorización — respuesta canónica del API."
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "PENDIENTE",
          "APROBADO_AUTO",
          "DOCS_PEDIDOS",
          "ESCALADO",
          "DECIDIDO"
        ],
        "title": "CaseStatus",
        "description": "CaseStatus — estado del workflow del caso (dashboard)."
      },
      "CaseSubmitIn": {
        "properties": {
          "report": {
            "$ref": "#/components/schemas/MedicalReportIn"
          },
          "policyNumber": {
            "type": "string",
            "title": "Policynumber",
            "description": "Número de la póliza del paciente."
          },
          "scenarioKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenariokey",
            "description": "Override opcional para forzar un escenario en demos."
          }
        },
        "type": "object",
        "required": [
          "report",
          "policyNumber"
        ],
        "title": "CaseSubmitIn",
        "description": "Body de `POST /api/v1/cases` — pedir autorización."
      },
      "CaseTraceOut": {
        "properties": {
          "trace": {
            "items": {
              "$ref": "#/components/schemas/TraceStepOut"
            },
            "type": "array",
            "title": "Trace"
          }
        },
        "type": "object",
        "required": [
          "trace"
        ],
        "title": "CaseTraceOut",
        "description": "Wrapper para `GET /api/v1/cases/{id}/trace`."
      },
...
      "DecidedBy": {
        "type": "string",
        "enum": [
          "agent",
          "auditor"
        ],
        "title": "DecidedBy",
        "description": "DecidedBy — quién emitió la decisión."
      },
      "DecisionOut": {
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/Outcome"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "decidedBy": {
            "$ref": "#/components/schemas/DecidedBy"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "type": "array",
            "title": "Evidence"
          },
          "missingDocs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missingdocs"
          },
          "escalationReason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EscalationReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "modelUsed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelused"
          }
        },
        "type": "object",
        "required": [
          "outcome",
          "rationale",
          "confidence",
          "decidedBy"
        ],
        "title": "DecisionOut",
        "description": "Decisión emitida por el agente o el auditor."
      },
...
      "EvidenceOut": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/EvidenceSource"
          },
          "field": {
            "type": "string",
            "title": "Field"
          },
          "quote": {
            "type": "string",
            "title": "Quote"
          }
        },
        "type": "object",
        "required": [
          "source",
          "field",
          "quote"
        ],
        "title": "EvidenceOut",
        "description": "Cita auditable que respalda la decisión."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MedicalReportIn": {
        "properties": {
          "patientId": {
            "type": "string",
            "title": "Patientid",
            "description": "ID del paciente (Pacientes:id)."
          },
          "format": {
            "$ref": "#/components/schemas/ReportFormat",
            "default": "text"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Contenido del informe (texto o ruta del PDF)."
          },
          "procedureSolicitedHint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proceduresolicitedhint",
            "description": "Procedimiento sugerido por el hospital (código o descripción)."
          },
          "diagnosis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "attendingDoctor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attendingdoctor"
          }
        },
        "type": "object",
        "required": [
          "patientId",
          "content"
        ],
        "title": "MedicalReportIn",
        "description": "Informe médico que el hospital adjunta al crear un caso."
      },
      "Outcome": {
        "type": "string",
        "enum": [
          "APPROVED_AUTO",
          "DOCS_REQUESTED",
          "ESCALATED",
          "DECIDED",
          "PENDING"
        ],
        "title": "Outcome",
        "description": "Outcome — resultado emitido por el agente o el auditor."
      },
...
      "ReportFormat": {
        "type": "string",
        "enum": [
          "text",
          "pdf"
        ],
        "title": "ReportFormat",
        "description": "ReportFormat — formato del informe médico."
      },
      "ResolveCaseIn": {
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED"
            ],
            "title": "Outcome"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mensaje humano final que verá el hospital."
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning",
            "description": "Razonamiento del auditor (rationale)."
          }
        },
        "type": "object",
        "required": [
          "outcome",
          "message",
          "reasoning"
        ],
        "title": "ResolveCaseIn",
        "description": "Body de `POST /api/v1/cases/{id}/resolve` — auditor cierra el caso."
      },
...
      "TraceStepOut": {
        "properties": {
          "node": {
            "type": "string",
            "title": "Node"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "state": {
            "$ref": "#/components/schemas/TraceStepState"
          },
          "durationMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationms"
          },
          "modelUsed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelused"
          },
          "tokensIn": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokensin"
          },
          "tokensOut": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokensout"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "node",
          "timestamp",
          "state"
        ],
        "title": "TraceStepOut",
        "description": "Paso individual de la traza del agente."
      },
      "TraceStepState": {
        "type": "string",
        "enum": [
          "running",
          "done",
          "error"
        ],
        "title": "TraceStepState",
        "description": "TraceStepState — estado terminal o en progreso de un paso de traza."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}